+++ src/de/lib_validate.rs
@@ -85,7 +85,7 @@ where
 
     enforce_single_document_and_finish(
         &mut src,
-        "use from_multiple or from_multiple_with_options",
+        "use from_str_multiple or from_str_multiple_with_options",
         wrap_err,
     )?;