+++ tests/serde_yaml/test_stream_deserializer.rs
@@ -9,7 +9,7 @@ struct Point {
 #[test]
 fn test_stream_deserializer() -> anyhow::Result<()> {
     let yaml = indoc!("---\nx: 1\n---\nx: 2\n");
-    let mut stream: std::vec::IntoIter<Point> = serde_saphyr::from_multiple(yaml)?.into_iter();
+    let mut stream: std::vec::IntoIter<Point> = serde_saphyr::from_str_multiple(yaml)?.into_iter();
     assert_eq!(stream.next().unwrap(), Point { x: 1 });
     assert_eq!(stream.next().unwrap(), Point { x: 2 });
     assert!(stream.next().is_none());