+++ tests/test_suite/y_U9NS.rs
@@ -16,7 +16,7 @@ fn yaml_u9ns_multi_documents() {
 time: 20:03:20\nplayer: Sammy Sosa\naction: strike (miss)\n...\n---\n
 time: 20:03:47\nplayer: Sammy Sosa\naction: grand slam\n...\n";
 
-    let docs: Vec<Play> = serde_saphyr::from_multiple(y).expect("failed to parse U9NS");
+    let docs: Vec<Play> = serde_saphyr::from_str_multiple(y).expect("failed to parse U9NS");
     assert_eq!(docs.len(), 2);
 
     assert_eq!(docs[0].time, "20:03:20");